Extract deployment addresses file content generation
